Fix monorepo problems
