+++ .eslintignore
@@ -1,4 +1,4 @@
 node_modules
 dist
 coverage
-examples/web-and-react-native/native/resolve
+examples/react-native-todo/native/resolve
