+++ examples/react-native-todo/native/client/containers/App.js
@@ -14,7 +14,7 @@ import {
   Platform
 } from 'react-native'
 
-import Item from '../components/Item'
+import Todo from '../components/Todo'
 
 const isAndroid = Platform.OS === 'android'
 const viewPadding = 10
@@ -23,6 +23,12 @@ const viewModelName = 'Todos'
 const aggregateId = 'root-id'
 
 const styles = StyleSheet.create({
+  layout: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center'
+  },
   container: {
     flex: 1,
     justifyContent: 'center',
@@ -51,7 +57,7 @@ const styles = StyleSheet.create({
   }
 })
 
-export class Todos extends React.PureComponent {
+export class App extends React.PureComponent {
   state = {
     text: ''
   }
@@ -91,7 +97,7 @@ export class Todos extends React.PureComponent {
   keyExtractor = item => item.id
 
   renderItem = ({ item: { checked, text, id } }) => (
-    <Item
+    <Todo
       checked={checked}
       text={text}
       removeItem={this.removeItem.bind(this, id)}
@@ -101,25 +107,27 @@ export class Todos extends React.PureComponent {
 
   render() {
     return (
-      <View
-        style={[styles.container, { paddingBottom: this.state.viewPadding }]}
-      >
-        <Text style={styles.title}>Tasks List</Text>
-        <FlatList
-          style={styles.list}
-          data={this.props.todos}
-          keyExtractor={this.keyExtractor}
-          renderItem={this.renderItem}
-        />
-        <TextInput
-          style={styles.textInput}
-          onChangeText={this.updateText}
-          onSubmitEditing={this.createItem}
-          value={this.state.text}
-          placeholder="New Task"
-          returnKeyType="done"
-          returnKeyLabel="done"
-        />
+      <View style={styles.layout}>
+        <View
+          style={[styles.container, { paddingBottom: this.state.viewPadding }]}
+        >
+          <Text style={styles.title}>Tasks List</Text>
+          <FlatList
+            style={styles.list}
+            data={this.props.todos}
+            keyExtractor={this.keyExtractor}
+            renderItem={this.renderItem}
+          />
+          <TextInput
+            style={styles.textInput}
+            onChangeText={this.updateText}
+            onSubmitEditing={this.createItem}
+            value={this.state.text}
+            placeholder="New Task"
+            returnKeyType="done"
+            returnKeyLabel="done"
+          />
+        </View>
       </View>
     )
   }
@@ -141,5 +149,5 @@ export default connectViewModel(mapStateToOptions)(
   connect(
     mapStateToProps,
     mapDispatchToProps
-  )(Todos)
+  )(App)
 )
