+++ examples/react-native-todo/native/jest.config.js
@@ -1,4 +1,11 @@
+const path = require('path')
+
 module.exports = {
+  rootDir: process.cwd(),
   testEnvironment: 'node',
-  preset: 'jest-expo'
+  preset: 'jest-expo',
+  setupFiles: ['raf/polyfill', '<rootDir>/jest-setup.js'],
+  transform: {
+    '^.+\\.js$': path.resolve(__dirname, 'jest.transform.js')
+  }
 }
