+++ examples/react-native-todo/native/jest.transform.js
@@ -1,0 +1,12 @@
+const babelrc = {
+  babelrc: false,
+  presets: ['@babel/preset-env', '@babel/preset-react', '@babel/preset-flow'],
+  plugins: [
+    '@babel/plugin-proposal-class-properties',
+    '@babel/plugin-proposal-export-default-from',
+    '@babel/plugin-proposal-export-namespace-from',
+    '@babel/plugin-transform-runtime'
+  ]
+}
+
+module.exports = require('babel-jest').createTransformer(babelrc)
