+++ examples/react-native-todo/web/package.json
@@ -30,7 +30,7 @@
     "babel-core": "^7.0.0-0",
     "babel-jest": "23.4.2",
     "jest": "23.4.2",
-    "my-local-ip": "^1.0.0",
+    "my-local-ip": "1.0.0",
     "react-test-renderer": "16.4.1",
     "regenerator-runtime": "0.12.0",
     "testcafe": "0.20.4"
